refactor(dynamodb): split helpers.rs by concerns (3309 -> 784 mod + 10 submodules)
Audit 2026-05-19 P9 remainder. 109-fn helpers.rs split:

- partiql.rs            (29 fns: tokenizer + parser + evaluator for PartiQL
                         where clauses + literal/object parsing)
- conditions.rs         (14 fns: condition expr eval + key cond + filter expr +
                         compare/size + split helpers + keyword match)
- schemas.rs             (9 fns: parse_key_schema / parse_attribute_definitions /
                         parse_provisioned_throughput / parse_gsi[_throughput] /
                         parse_lsi / parse_tags / expression-attr-name+value parse)
- paths.rs               (9 fns: project_item / resolve_projection_path /
                         resolve_nested_path / parse_path_segments /
                         insert_nested_value / wrap_value_in_path /
                         merge_attribute_values / resolve_path / resolve_attr_name)
- table_lookup.rs        (7 fns: get_table[_mut][_with_code] / find_table_by_arn[_mut] /
                         resolve_table_name)
- keys.rs                (6 fns: extract_key / parse_key_map / item_matches_key /
                         validate_key_in_item / validate_key_attributes_in_key /
                         extract_key_for_schema)
- updates.rs             (5 fns: apply_add/delete_assignment / parse_arithmetic /
                         extract_number / assign_nested_path)
- request.rs             (4 fns: is_mutating_action / require_str[_with_code] /
                         require_object)
- metrics.rs             (4 fns: build_consumed_capacity / return_consumed_mode /
                         return_icm_mode / build_item_collection_metrics)
- table_descriptions.rs  (2 fns: build_table_description{,_json})

Each family re-exported pub(crate) from mod.rs so existing callers keep
their bare-fn-name imports.

272 unit tests pass.
