+++ crates/fakecloud-dynamodb/src/service/helpers/request.rs
@@ -1,0 +1,84 @@
+//! dynamodb helpers `request` concerns (audit-2026-05-19).
+
+use super::*;
+
+/// Actions that mutate DynamoDB state and therefore require a snapshot
+/// write after success. Kept in sync with the dispatch table above.
+pub(crate) fn is_mutating_action(action: &str) -> bool {
+    matches!(
+        action,
+        "CreateTable"
+            | "DeleteTable"
+            | "UpdateTable"
+            | "PutItem"
+            | "DeleteItem"
+            | "UpdateItem"
+            | "BatchWriteItem"
+            | "TagResource"
+            | "UntagResource"
+            | "TransactWriteItems"
+            | "ExecuteStatement"
+            | "BatchExecuteStatement"
+            | "ExecuteTransaction"
+            | "UpdateTimeToLive"
+            | "PutResourcePolicy"
+            | "DeleteResourcePolicy"
+            | "CreateBackup"
+            | "DeleteBackup"
+            | "RestoreTableFromBackup"
+            | "RestoreTableToPointInTime"
+            | "UpdateContinuousBackups"
+            | "CreateGlobalTable"
+            | "UpdateGlobalTable"
+            | "UpdateGlobalTableSettings"
+            | "UpdateTableReplicaAutoScaling"
+            | "EnableKinesisStreamingDestination"
+            | "DisableKinesisStreamingDestination"
+            | "UpdateKinesisStreamingDestination"
+            | "UpdateContributorInsights"
+            | "ExportTableToPointInTime"
+            | "ImportTable"
+    )
+}
+
+// ── Helper functions ────────────────────────────────────────────────────
+
+pub(crate) fn require_str<'a>(body: &'a Value, field: &str) -> Result<&'a str, AwsServiceError> {
+    require_str_with_code(body, field, "ValidationException")
+}
+
+/// Like [`require_str`] but emits a caller-chosen wire error code.
+///
+/// Several DynamoDB ops (e.g. CreateBackup, DescribeContinuousBackups,
+/// UpdateContinuousBackups, ExportTableToPointInTime,
+/// RestoreTableToPointInTime, ListBackups, ListImports) do not declare
+/// `ValidationException` in their Smithy `errors:` list. Use this helper
+/// to surface a declared shape (e.g. `TableNotFoundException`) for the
+/// missing-required-field path on those ops.
+pub(crate) fn require_str_with_code<'a>(
+    body: &'a Value,
+    field: &str,
+    code: &str,
+) -> Result<&'a str, AwsServiceError> {
+    body[field].as_str().ok_or_else(|| {
+        AwsServiceError::aws_error(
+            StatusCode::BAD_REQUEST,
+            code,
+            format!("{field} is required"),
+        )
+    })
+}
+
+pub(crate) fn require_object(
+    body: &Value,
+    field: &str,
+) -> Result<HashMap<String, AttributeValue>, AwsServiceError> {
+    let obj = body[field].as_object().ok_or_else(|| {
+        AwsServiceError::aws_error(
+            StatusCode::BAD_REQUEST,
+            "ValidationException",
+            format!("{field} is required"),
+        )
+    })?;
+    Ok(obj.iter().map(|(k, v)| (k.clone(), v.clone())).collect())
+}
