+++ crates/fakecloud-dynamodb/src/service/helpers/updates.rs
@@ -1,0 +1,199 @@
+//! dynamodb helpers `updates` concerns (audit-2026-05-19).
+
+use super::*;
+
+/// Write `value` at a dotted path inside an M-typed attribute.
+///
+/// Resolves each `#name` segment through `expr_attr_names`. The top-level
+/// attribute and every intermediate segment must already exist as a Map —
+/// DynamoDB rejects writes through missing parents with ValidationException.
+pub(crate) fn assign_nested_path(
+    item: &mut HashMap<String, AttributeValue>,
+    path: &str,
+    expr_attr_names: &HashMap<String, String>,
+    value: Value,
+) -> Result<(), AwsServiceError> {
+    let mut segments: Vec<String> = path
+        .split('.')
+        .map(|seg| resolve_attr_name(seg.trim(), expr_attr_names))
+        .collect();
+    if segments.len() < 2 {
+        return Err(invalid_document_path());
+    }
+
+    let leaf = segments.pop().expect("len >= 2");
+    let top = segments.remove(0);
+
+    let top_attr = item.get_mut(&top).ok_or_else(invalid_document_path)?;
+    let mut current = top_attr
+        .get_mut("M")
+        .and_then(|m| m.as_object_mut())
+        .ok_or_else(invalid_document_path)?;
+
+    for seg in &segments {
+        current = current
+            .get_mut(seg)
+            .and_then(|v| v.get_mut("M"))
+            .and_then(|m| m.as_object_mut())
+            .ok_or_else(invalid_document_path)?;
+    }
+
+    current.insert(leaf, value);
+    Ok(())
+}
+
+pub(crate) fn extract_number(val: &Option<Value>) -> Option<f64> {
+    val.as_ref()
+        .and_then(|v| v.get("N"))
+        .and_then(|n| n.as_str())
+        .and_then(|s| s.parse().ok())
+}
+
+pub(crate) fn parse_arithmetic(expr: &str) -> Option<(&str, &str, bool)> {
+    let mut depth = 0;
+    for (i, c) in expr.char_indices() {
+        match c {
+            '(' => depth += 1,
+            ')' => depth -= 1,
+            '+' if depth == 0 && i > 0 => {
+                return Some((&expr[..i], &expr[i + 1..], true));
+            }
+            '-' if depth == 0 && i > 0 => {
+                return Some((&expr[..i], &expr[i + 1..], false));
+            }
+            _ => {}
+        }
+    }
+    None
+}
+
+pub(crate) fn apply_add_assignment(
+    item: &mut HashMap<String, AttributeValue>,
+    assignment: &str,
+    expr_attr_names: &HashMap<String, String>,
+    expr_attr_values: &HashMap<String, Value>,
+) -> Result<(), AwsServiceError> {
+    let parts: Vec<&str> = assignment.splitn(2, ' ').collect();
+    if parts.len() != 2 {
+        return Ok(());
+    }
+
+    let attr = resolve_attr_name(parts[0].trim(), expr_attr_names);
+    let val_ref = parts[1].trim();
+    let add_val = expr_attr_values.get(val_ref);
+
+    if let Some(add_val) = add_val {
+        if let Some(existing) = item.get(&attr) {
+            if let (Some(existing_num), Some(add_num)) = (
+                extract_number(&Some(existing.clone())),
+                extract_number(&Some(add_val.clone())),
+            ) {
+                let result = existing_num + add_num;
+                let num_str = if result == result.trunc() {
+                    format!("{}", result as i64)
+                } else {
+                    format!("{result}")
+                };
+                item.insert(attr, json!({"N": num_str}));
+            } else if let Some(existing_set) = existing.get("SS").and_then(|v| v.as_array()) {
+                if let Some(add_set) = add_val.get("SS").and_then(|v| v.as_array()) {
+                    let mut merged: Vec<Value> = existing_set.clone();
+                    for v in add_set {
+                        if !merged.contains(v) {
+                            merged.push(v.clone());
+                        }
+                    }
+                    item.insert(attr, json!({"SS": merged}));
+                }
+            } else if let Some(existing_set) = existing.get("NS").and_then(|v| v.as_array()) {
+                if let Some(add_set) = add_val.get("NS").and_then(|v| v.as_array()) {
+                    let mut merged: Vec<Value> = existing_set.clone();
+                    for v in add_set {
+                        if !merged.contains(v) {
+                            merged.push(v.clone());
+                        }
+                    }
+                    item.insert(attr, json!({"NS": merged}));
+                }
+            } else if let Some(existing_set) = existing.get("BS").and_then(|v| v.as_array()) {
+                if let Some(add_set) = add_val.get("BS").and_then(|v| v.as_array()) {
+                    let mut merged: Vec<Value> = existing_set.clone();
+                    for v in add_set {
+                        if !merged.contains(v) {
+                            merged.push(v.clone());
+                        }
+                    }
+                    item.insert(attr, json!({"BS": merged}));
+                }
+            }
+        } else {
+            item.insert(attr, add_val.clone());
+        }
+    }
+
+    Ok(())
+}
+
+pub(crate) fn apply_delete_assignment(
+    item: &mut HashMap<String, AttributeValue>,
+    assignment: &str,
+    expr_attr_names: &HashMap<String, String>,
+    expr_attr_values: &HashMap<String, Value>,
+) -> Result<(), AwsServiceError> {
+    let parts: Vec<&str> = assignment.splitn(2, ' ').collect();
+    if parts.len() != 2 {
+        return Ok(());
+    }
+
+    let attr = resolve_attr_name(parts[0].trim(), expr_attr_names);
+    let val_ref = parts[1].trim();
+    let del_val = expr_attr_values.get(val_ref);
+
+    if let (Some(existing), Some(del_val)) = (item.get(&attr).cloned(), del_val) {
+        if let (Some(existing_set), Some(del_set)) = (
+            existing.get("SS").and_then(|v| v.as_array()),
+            del_val.get("SS").and_then(|v| v.as_array()),
+        ) {
+            let filtered: Vec<Value> = existing_set
+                .iter()
+                .filter(|v| !del_set.contains(v))
+                .cloned()
+                .collect();
+            if filtered.is_empty() {
+                item.remove(&attr);
+            } else {
+                item.insert(attr, json!({"SS": filtered}));
+            }
+        } else if let (Some(existing_set), Some(del_set)) = (
+            existing.get("NS").and_then(|v| v.as_array()),
+            del_val.get("NS").and_then(|v| v.as_array()),
+        ) {
+            let filtered: Vec<Value> = existing_set
+                .iter()
+                .filter(|v| !del_set.contains(v))
+                .cloned()
+                .collect();
+            if filtered.is_empty() {
+                item.remove(&attr);
+            } else {
+                item.insert(attr, json!({"NS": filtered}));
+            }
+        } else if let (Some(existing_set), Some(del_set)) = (
+            existing.get("BS").and_then(|v| v.as_array()),
+            del_val.get("BS").and_then(|v| v.as_array()),
+        ) {
+            let filtered: Vec<Value> = existing_set
+                .iter()
+                .filter(|v| !del_set.contains(v))
+                .cloned()
+                .collect();
+            if filtered.is_empty() {
+                item.remove(&attr);
+            } else {
+                item.insert(attr, json!({"BS": filtered}));
+            }
+        }
+    }
+
+    Ok(())
+}
